writing scripts

+++ testAPI/scripts/execute_job.sh
@@ -1,10 +1,9 @@
 #!/bin/bash
-
-node=$1
-scheduler_host=$2
-rm -rf ${node}-l.log
-touch ${node}-l.log
-scheduler_port="9000"
+usage() {
+    echo "usage: execute_job.sh [node scheduler_host scheduler_port mode pattern]"
+    echo "mode: CPU, Local, TCP, RDMA, Hybrid"
+    echo "pattern: Large Median Small Mixed"
+}
 
 processLine(){
 	line="$@" # get all args
@@ -13,30 +12,45 @@ processLine(){
     arg3=$(echo $line | awk '{ print $3 }')
 
 	sleep $arg3
-	/home/tian/testAPI/test_bench $arg1 $arg2 $scheduler_host $scheduler_port & 
+	./job-testbench.sh $arg1 $arg2 $scheduler_host $scheduler_port $mode & 
 }
 
+if [[ "$#" -eq "4" ]]; then
+    node=$1
+    scheduler_host=$2
+    scheduler_port=$3
+    mode=$4         #CPU, Local, TCP, FPGA, Hybrid
+    pattern=$5      #Large, Median, Small, Mixed
+    
+    write_log="logfile/${node}-${mode}-${pattern}.log"
+    rm -rf $write_log 
+    touch $write_log
+    
+    FILE="job-pattern/$node-$pattern.txt"
+    
+    if [ ! -f $FILE ]; then
+    	echo "$FILE: does not exists"
+    	exit 1
+    elif [ ! -r $FILE ]; then
+    	echo "$FILE: can not be read"
+    	exit 2
+    fi
 
-FILE="job-$node.txt"
-if [ ! -f $FILE ]; then
-	echo "$FILE: does not exists"
-	exit 1
-elif [ ! -r $FILE ]; then
-	echo "$FILE: can not be read"
-	exit 2
+    BAKIFS=$IFS
+    IFS=$(echo -en "\n\b")
+    exec 3<&0
+    exec 0<"$FILE"
+    SHELLPID=$$
+    while read -r line
+    do
+    	processLine $line >>$write_log
+    	#processline $line
+    done
+    exec 0<&3
+    IFS=$BAKIFS
+    kill $SHELLPID
+else
+    usage
 fi
 
-BAKIFS=$IFS
-IFS=$(echo -en "\n\b")
-exec 3<&0
-exec 0<"$FILE"
-SHELLPID=$$
-while read -r line
-do
-	processLine $line >>${node}-l.log
-	#processLine $line
-done
-exec 0<&3
-IFS=$BAKIFS
-kill $SHELLPID
-exit 0
+exit
